refactor code to be easier to repurpose
