+++ .gitignore
@@ -1,8 +1,7 @@
 # Ignore virtual environment directories
 venv/
 # Ignore the local files
-assistant_serapi_key.txt
-assistant.sh
+settings.py
 # Ignore piper
 piper*
 cori*
