+++ assistant.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+source /home/bjoern/VM/NOVA_AIDAS/venv/bin/activate
+python3 /home/bjoern/VM/NOVA_AIDAS/assistant_local.py /home/bjoern/VM/NOVA_AIDAS
